requirements.txt
certifi==2018.11.29
chardet==3.0.4
gevent==1.3.7
googlemaps==3.0.2
greenlet==0.4.15
gspread==3.1.0
httplib2==0.12.0
idna==2.8
oauth2client==4.1.3
pushbullet.py==0.11.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
weather-api==1.0.6
websocket-client==0.54.0
